+++ packages/pangea-node-sdk/src/index.ts
@@ -29,6 +29,7 @@ export const PangeaResponse = _PangeaResponse;
 
 export const AuditService = services.AuditService;
 export const AuthNService = services.AuthNService;
+export const AuthZService = services.AuthZService;
 export const BaseService = services.BaseService;
 export const EmbargoService = services.EmbargoService;
 export const RedactService = services.RedactService;