+++ packages/pangea-node-sdk/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Documentation to service client constructors.
 - Support for Secure Audit Log's log stream API.
 - Support for Secure Audit Log's export API.
+- AuthZ service support.
 
 ### Fixed